requirements.txt
pyTelegramBotAPI
nltk
scikit-learn
pandas
joblib
python-docx
openpyxl